// MCP API/gen/cmp/types/v4/date_pb.ts
// @generated by protoc-gen-es v2.10.1 with parameter "target=ts,import_extension=.js"
// @generated from file cmp/types/v4/date.proto (package cmp.types.v4, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import { file_buf_validate_validate } from "../../../buf/validate/validate_pb.js";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file cmp/types/v4/date.proto.
 */
export const file_cmp_types_v4_date: GenFile = /*@__PURE__*/
  fileDesc("ChdjbXAvdHlwZXMvdjQvZGF0ZS5wcm90bxIMY21wLnR5cGVzLnY0IlAKBERhdGUSFgoEeWVhchgBIAEoDUIIukgFKgMYj04SGAoFbW9udGgYAiABKA1CCbpIBioEGAwoARIWCgNkYXkYAyABKA1CCbpIBioEGB8oAWIGcHJvdG8z", [file_buf_validate_validate]);

/**
 * Represents a calendar date, specified by year, month, and day.
 *
 * This message is used for scenarios where only the date component is needed,
 * without time or timezone information.
 *
 * ![Diagram](https://storage.googleapis.com/docs-cmp-files/diagrams/c4t/proto/cmp/types/v4/date.proto.dot.xs.svg)
 *
 * [Open Message Diagram](https://storage.googleapis.com/docs-cmp-files/diagrams/c4t/proto/cmp/types/v4/date.proto.dot.svg)
 *
 * @generated from message cmp.types.v4.Date
 */
export type Date = Message<"cmp.types.v4.Date"> & {
  /**
   * The year of the date. For example, 2025.
   *
   * @generated from field: uint32 year = 1;
   */
  year: number;

  /**
   * The month of the year, from 1 (January) to 12 (December).
   *
   * @generated from field: uint32 month = 2;
   */
  month: number;

  /**
   * The day of the month, from 1 to 31.
   *
   * The validity of the day for a given month and year (e.g., ensuring February
   * doesn't have 30 days) should be handled by application logic.
   *
   * @generated from field: uint32 day = 3;
   */
  day: number;
};

/**
 * Describes the message cmp.types.v4.Date.
 * Use `create(DateSchema)` to create a new message.
 */
export const DateSchema: GenMessage<Date> = /*@__PURE__*/
  messageDesc(file_cmp_types_v4_date, 0);